initial navigation structure
